Rely fully on go-swagger's email format validator
go-swagger provides a number of format validators out of the box, we were rolling our own email formatter but I'd rather we use the built in one. Since it is built in, that means that we don't need to provide our own pattern in internal.yaml and need to put the validator into json schema field instead.

In fixing this I noticed that we were not correctly taking advantage of the builtin validator for SSN either, so I've fixed that as well.